feat: Next.js App Routerへの移行とsrc/ディレクトリ削除
- Pages RouterからApp Routerに移行
- src/ディレクトリを削除し、ルート直下に再配置
  - src/components → components/
  - src/types → types/
  - src/utils + src/constants → lib/
  - src/styles → styles/
- tsconfig.jsonとvitest.config.tsのパス設定を更新
- Chakra UI v3のApp Router対応
- 全テスト通過とビルド成功を確認
