+++ components/page/index/BlogAccounts.tsx
@@ -42,7 +42,7 @@ export default BlogAccounts;
 if (import.meta.vitest) {
 	const { describe, expect, it } = import.meta.vitest;
 	const { screen } = await import("@testing-library/react");
-	const { renderWithChakra } = await import("@/utils/test-utils");
+	const { renderWithChakra } = await import("@/lib/test-utils");
 
 	describe("BlogAccounts component", () => {
 		it("renders blog accounts correctly", () => {
