+++ components/page/index/profile.tsx
@@ -71,7 +71,7 @@ export default Profile;
 if (import.meta.vitest) {
 	const { describe, expect, it } = import.meta.vitest;
 	const { screen } = await import("@testing-library/react");
-	const { renderWithChakra } = await import("@/utils/test-utils");
+	const { renderWithChakra } = await import("@/lib/test-utils");
 
 	describe("Profile component", () => {
 		it("renders correctly", () => {
